# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/cluster-create": {
            "post": {
                "tags": [
                    "私有方法"
                ],
                "summary": "集群创建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "master_url",
                        "name": "master_url",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "config",
                        "name": "config",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/cluster-node": {
            "get": {
                "tags": [
                    "私有方法"
                ],
                "summary": "获取node节点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "tags": [
                    "测试方法"
                ],
                "summary": "ping",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}